1.修复删除回复无权限的bug 2.修正加入热帖榜的条件，少于10个时可以直接加入 3.修正缓存为空时出错 4.修正热帖榜和index不统一 5.修复身份更换后其他人身份一起更换了（之前修改的是角色表，所以出了问题） 6.修复登录后的跳转页面问题 7.过滤敏感词，添加到敏感词表中 8.统计用户敏感词出现次数 9.用户删除帖子 完成 10.用户主页分页 完成 11.修复收藏和我的帖子的分页bug 完成 12.禁言 OA页面显示 完成 13.优化用户体验（自动刷新OA页面表格,优化回复定位功能） 14.修正主页和我的消息页面的链接跳转

+++ bbs-main/src/main/java/com/thorn/bbsmain/services/ReplyService.java
@@ -13,6 +13,7 @@ import com.thorn.bbsmain.mapper.entity.User;
 import com.thorn.bbsmain.utils.MsgBuilder;
 import com.thorn.bbsmain.utils.MyUtil;
 import com.thorn.bbsmain.utils.message.MessageObject;
+import com.thorn.bbsmain.utils.review.ReviewInfo;
 import impl.HotPointManager;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
@@ -228,7 +229,6 @@ public class ReplyService {
         //不给自己发消息
         if (postOwner != reply.getReplyer()) {
             addMessage(reply, postOwner);
-
         }
         if (reply.getReplyTo() != null) {
             int replyToID = replyMapper.getReplyOwner(reply.getReplyTo(), reply.getPostid());
@@ -236,6 +236,8 @@ public class ReplyService {
                 addMessage(reply, replyToID);
             }
         }
+        rabbitTemplate.convertAndSend("review",
+                ReviewInfo.builder().uid(reply.getReplyer()).pid(reply.getPostid()).floor(reply.getFloor()).content(reply.getContent()).build());
         return builder.getMsg("redirect:/post/" + reply.getPostid() + "/-1/" + reply.getId());
     }
 
@@ -266,7 +268,7 @@ public class ReplyService {
         if (!isExist(pid, floor)) {
             throw new DeleteReplyException("删除回复参数错误，不存在该帖子或者楼层");
         }
-        if (!userService.hasRole("admin") || replyMapper.hasPermission(user.getUid(), pid, floor) == 0) {
+        if (!userService.hasRole("admin") && replyMapper.hasPermission(user.getUid(), pid, floor) == 0) {
             throw new DeleteReplyException("没有权限删除");
         }
 
@@ -285,5 +287,17 @@ public class ReplyService {
 
     boolean isExist(int pid, int floor) {
         return replyMapper.isExist(pid, floor) > 0;
+    }
+
+    public int getReplyOffsetByFloor(int pid, int floor) {
+        return replyMapper.getOffsetByFloor(pid, floor);
+    }
+
+    public int getONE_PAGE_REPLY_NUM() {
+        return ONE_PAGE_REPLY_NUM;
+    }
+
+    public int getReplyIDByFloor(int pid, int floor) {
+        return replyMapper.getReplyIDByFloor(pid, floor);
     }
 }
